- fixed compilation warnings.

+++ src/textures/texturemanager.cpp
@@ -1450,7 +1450,7 @@ DEFINE_ACTION_FUNCTION_NATIVE(_TexMan, GetSize, GetTextureSize)
 static void GetScaledSize(int texid, DVector2 *pvec)
 {
 	auto tex = TexMan.ByIndex(texid);
-	int x, y;
+	double x, y;
 	if (tex != nullptr)
 	{
 		x = tex->GetScaledWidthDouble();
@@ -1481,7 +1481,7 @@ DEFINE_ACTION_FUNCTION_NATIVE(_TexMan, GetScaledSize, GetScaledSize)
 static void GetScaledOffset(int texid, DVector2 *pvec)
 {
 	auto tex = TexMan.ByIndex(texid);
-	int x, y;
+	double x, y;
 	if (tex != nullptr)
 	{
 		x = tex->GetScaledLeftOffsetDouble(0);
